+++ runtime/src/static/scala/DeliteInterfaces.scala
@@ -6,11 +6,11 @@ package generated.scala
  */
 
 abstract class DeliteOpMultiLoop[A] {
-  def size: Int
-  var loopStart: Int
-  var loopSize: Int
+  def size: Long
+  var loopStart: Long
+  var loopSize: Long
   def alloc: A
-  def processRange(__act: A, start: Int, end: Int): A //init+process
+  def processRange(__act: A, start: Long, end: Long): A //init+process
   def combine(__act: A, rhs: A): Unit
   def postCombine(__act: A, rhs: A): Unit
   def postProcInit(__act: A): Unit
